Factor public type definitions out of services.ts into services/types.ts
