+++ Jakefile.js
@@ -102,6 +102,7 @@ var servicesSources = [
     "services.ts",
     "shims.ts",
     "signatureHelp.ts",
+    "types.ts",
     "utilities.ts",
     "formatting/formatting.ts",
     "formatting/formattingContext.ts",
