+++ src/services/tsconfig.json
@@ -35,6 +35,7 @@
         "services.ts",
         "shims.ts",
         "signatureHelp.ts",
+        "types.ts",
         "utilities.ts",
         "jsTyping.ts",
         "formatting/formatting.ts",
